requirements.txt
pytesseract
Pillow
numpy
gTTS
moviepy

googletrans==4.0.0-rc1 
edge-tts